change:  - moved actions from networkAction to earnAction  - make use of type LiquidityPoolLayer instead of EpoolLayer from network service cleanup.
