+++ src/actions/networkAction.ts
@@ -13,7 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License. */
 
-import networkService, { EPoolLayer } from 'services/networkService'
+import networkService from 'services/networkService'
 import transactionService from 'services/transaction.service'
 import { createAction } from './createAction'
 import { BigNumberish } from 'ethers'
@@ -74,29 +74,6 @@ export const depositETHL2 = (payload) =>
 //DEPOSIT ERC20
 export const depositErc20 = (payload) =>
   createAction('DEPOSIT/CREATE', () => networkService.depositErc20(payload))
-
-// FIXME: move this earn related actions.
-export const getReward = (
-  currencyAddress: string,
-  value_Wei_String: BigNumberish,
-  L1orL2Pool: EPoolLayer
-) =>
-  createAction('EARN/HARVEST', () =>
-    networkService.getReward(currencyAddress, value_Wei_String, L1orL2Pool)
-  )
-
-export const withdrawLiquidity = (
-  currencyAddress: string,
-  value_Wei_String: string,
-  L1orL2Pool: EPoolLayer
-) =>
-  createAction('EARN/WITHDRAW', () =>
-    networkService.withdrawLiquidity(
-      currencyAddress,
-      value_Wei_String,
-      L1orL2Pool
-    )
-  )
 
 export const approveERC20 = (
   value: BigNumberish,
